Revert "Protect all SingleResultCallback observables from NPE"
This reverts commit 3c610903

JAVARS-220
