+++ driver/src/main/com/mongodb/reactivestreams/client/internal/GridFSBucketImpl.java
@@ -165,13 +165,13 @@ public final class GridFSBucketImpl implements GridFSBucket {
 
     @Override
     public Publisher<ObjectId> uploadFromStream(final String filename, final AsyncInputStream source, final GridFSUploadOptions options) {
-        return new SingleResultObservableToPublisher<ObjectId>(
+        return new ObservableToPublisher<ObjectId>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<ObjectId>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<ObjectId> callback) {
                         wrapped.uploadFromStream(filename, toCallbackAsyncInputStream(source), options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -182,14 +182,14 @@ public final class GridFSBucketImpl implements GridFSBucket {
     @Override
     public Publisher<Success> uploadFromStream(final BsonValue id, final String filename, final AsyncInputStream source,
                                                final GridFSUploadOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.uploadFromStream(id, filename, toCallbackAsyncInputStream(source), options,
                                 voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -200,14 +200,14 @@ public final class GridFSBucketImpl implements GridFSBucket {
     @Override
     public Publisher<ObjectId> uploadFromStream(final ClientSession clientSession, final String filename, final AsyncInputStream source,
                                                 final GridFSUploadOptions options) {
-        return new SingleResultObservableToPublisher<ObjectId>(
+        return new ObservableToPublisher<ObjectId>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<ObjectId>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<ObjectId> callback) {
                         wrapped.uploadFromStream(clientSession.getWrapped(), filename, toCallbackAsyncInputStream(source), options,
                                 callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -219,14 +219,14 @@ public final class GridFSBucketImpl implements GridFSBucket {
     @Override
     public Publisher<Success> uploadFromStream(final ClientSession clientSession, final BsonValue id, final String filename,
                                                final AsyncInputStream source, final GridFSUploadOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.uploadFromStream(clientSession.getWrapped(), id, filename, toCallbackAsyncInputStream(source), options,
                                 voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -272,25 +272,25 @@ public final class GridFSBucketImpl implements GridFSBucket {
 
     @Override
     public Publisher<Long> downloadToStream(final ObjectId id, final AsyncOutputStream destination) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(id, toCallbackAsyncOutputStream(destination), callback);
                     }
-                });
+                }));
     }
 
 
     @Override
     public Publisher<Long> downloadToStream(final BsonValue id, final AsyncOutputStream destination) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(id, toCallbackAsyncOutputStream(destination), callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -301,35 +301,35 @@ public final class GridFSBucketImpl implements GridFSBucket {
     @Override
     public Publisher<Long> downloadToStream(final String filename, final AsyncOutputStream destination,
                                             final GridFSDownloadOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(filename, toCallbackAsyncOutputStream(destination), options, callback);
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Long> downloadToStream(final ClientSession clientSession, final ObjectId id, final AsyncOutputStream destination) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(clientSession.getWrapped(), id, toCallbackAsyncOutputStream(destination), callback);
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Long> downloadToStream(final ClientSession clientSession, final BsonValue id, final AsyncOutputStream destination) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(clientSession.getWrapped(), id, toCallbackAsyncOutputStream(destination), callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -341,14 +341,14 @@ public final class GridFSBucketImpl implements GridFSBucket {
     @Override
     public Publisher<Long> downloadToStream(final ClientSession clientSession, final String filename, final AsyncOutputStream destination,
                                             final GridFSDownloadOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.downloadToStream(clientSession.getWrapped(), filename, toCallbackAsyncOutputStream(destination), options,
                                 callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -373,112 +373,112 @@ public final class GridFSBucketImpl implements GridFSBucket {
 
     @Override
     public Publisher<Success> delete(final ObjectId id) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.delete(id, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> delete(final BsonValue id) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.delete(id, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> delete(final ClientSession clientSession, final ObjectId id) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.delete(clientSession.getWrapped(), id, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> delete(final ClientSession clientSession, final BsonValue id) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.delete(clientSession.getWrapped(), id, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> rename(final ObjectId id, final String newFilename) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.rename(id, newFilename, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> rename(final BsonValue id, final String newFilename) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.rename(id, newFilename, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> rename(final ClientSession clientSession, final ObjectId id, final String newFilename) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.rename(clientSession.getWrapped(), id, newFilename, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> rename(final ClientSession clientSession, final BsonValue id, final String newFilename) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.rename(clientSession.getWrapped(), id, newFilename, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> drop() {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.drop(voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> drop(final ClientSession clientSession) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.drop(clientSession.getWrapped(), voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
 }
