+++ driver/src/main/com/mongodb/reactivestreams/client/internal/ListIndexesPublisherImpl.java
@@ -50,13 +50,13 @@ final class ListIndexesPublisherImpl<TResult> implements ListIndexesPublisher<TR
 
     @Override
     public Publisher<TResult> first() {
-        return new SingleResultObservableToPublisher<TResult>(
+        return new ObservableToPublisher<TResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TResult> callback) {
                         wrapped.first(callback);
                     }
-                });
+                }));
     }
 
     @Override
