+++ driver/src/main/com/mongodb/reactivestreams/client/internal/MongoCollectionImpl.java
@@ -143,13 +143,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Deprecated
     @SuppressWarnings("deprecation")
     public Publisher<Long> count(final Bson filter, final CountOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.count(filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -168,13 +168,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Deprecated
     @SuppressWarnings("deprecation")
     public Publisher<Long> count(final ClientSession clientSession, final Bson filter, final CountOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.count(clientSession.getWrapped(), filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -184,13 +184,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Long> estimatedDocumentCount(final EstimatedDocumentCountOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.estimatedDocumentCount(options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -205,13 +205,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Long> countDocuments(final Bson filter, final CountOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.countDocuments(filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -226,13 +226,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Long> countDocuments(final ClientSession clientSession, final Bson filter, final CountOptions options) {
-        return new SingleResultObservableToPublisher<Long>(
+        return new ObservableToPublisher<Long>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Long>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Long> callback) {
                         wrapped.countDocuments(clientSession.getWrapped(), filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -390,13 +390,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<BulkWriteResult> bulkWrite(final List<? extends WriteModel<? extends TDocument>> requests,
                                                 final BulkWriteOptions options) {
-        return new SingleResultObservableToPublisher<BulkWriteResult>(
+        return new ObservableToPublisher<BulkWriteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<BulkWriteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<BulkWriteResult> callback) {
                         wrapped.bulkWrite(requests, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -409,13 +409,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     public Publisher<BulkWriteResult> bulkWrite(final ClientSession clientSession,
                                                 final List<? extends WriteModel<? extends TDocument>> requests,
                                                 final BulkWriteOptions options) {
-        return new SingleResultObservableToPublisher<BulkWriteResult>(
+        return new ObservableToPublisher<BulkWriteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<BulkWriteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<BulkWriteResult> callback) {
                         wrapped.bulkWrite(clientSession.getWrapped(), requests, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -425,13 +425,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Success> insertOne(final TDocument document, final InsertOneOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.insertOne(document, options, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -441,13 +441,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Success> insertOne(final ClientSession clientSession, final TDocument document, final InsertOneOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.insertOne(clientSession.getWrapped(), document, options, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -457,13 +457,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Success> insertMany(final List<? extends TDocument> documents, final InsertManyOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.insertMany(documents, options, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -474,13 +474,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<Success> insertMany(final ClientSession clientSession, final List<? extends TDocument> documents,
                                          final InsertManyOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.insertMany(clientSession.getWrapped(), documents, options, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -490,13 +490,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<DeleteResult> deleteOne(final Bson filter, final DeleteOptions options) {
-        return new SingleResultObservableToPublisher<DeleteResult>(
+        return new ObservableToPublisher<DeleteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<DeleteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<DeleteResult> callback) {
                         wrapped.deleteOne(filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -506,13 +506,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<DeleteResult> deleteOne(final ClientSession clientSession, final Bson filter, final DeleteOptions options) {
-        return new SingleResultObservableToPublisher<DeleteResult>(
+        return new ObservableToPublisher<DeleteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<DeleteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<DeleteResult> callback) {
                         wrapped.deleteOne(clientSession.getWrapped(), filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -522,13 +522,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<DeleteResult> deleteMany(final Bson filter, final DeleteOptions options) {
-        return new SingleResultObservableToPublisher<DeleteResult>(
+        return new ObservableToPublisher<DeleteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<DeleteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<DeleteResult> callback) {
                         wrapped.deleteMany(filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -538,13 +538,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<DeleteResult> deleteMany(final ClientSession clientSession, final Bson filter, final DeleteOptions options) {
-        return new SingleResultObservableToPublisher<DeleteResult>(
+        return new ObservableToPublisher<DeleteResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<DeleteResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<DeleteResult> callback) {
                         wrapped.deleteMany(clientSession.getWrapped(), filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -554,26 +554,26 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<UpdateResult> replaceOne(final Bson filter, final TDocument replacement, final ReplaceOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.replaceOne(filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
     @Deprecated
     public Publisher<UpdateResult> replaceOne(final Bson filter, final TDocument replacement, final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @SuppressWarnings("deprecation")
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.replaceOne(filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -584,27 +584,27 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<UpdateResult> replaceOne(final ClientSession clientSession, final Bson filter, final TDocument replacement,
                                               final ReplaceOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.replaceOne(clientSession.getWrapped(), filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
     @Deprecated
     public Publisher<UpdateResult> replaceOne(final ClientSession clientSession, final Bson filter, final TDocument replacement,
                                               final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @SuppressWarnings("deprecation")
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.replaceOne(clientSession.getWrapped(), filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -614,13 +614,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<UpdateResult> updateOne(final Bson filter, final Bson update, final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateOne(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -631,13 +631,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<UpdateResult> updateOne(final ClientSession clientSession, final Bson filter, final Bson update,
                                              final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateOne(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -647,13 +647,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<UpdateResult> updateOne(final Bson filter, final List<? extends Bson> update, final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateOne(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -664,13 +664,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<UpdateResult> updateOne(final ClientSession clientSession, final Bson filter, final List<? extends Bson> update,
                                              final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateOne(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -680,13 +680,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<UpdateResult> updateMany(final Bson filter, final Bson update, final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateMany(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -697,13 +697,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<UpdateResult> updateMany(final ClientSession clientSession, final Bson filter, final Bson update,
                                               final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateMany(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -713,13 +713,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<UpdateResult> updateMany(final Bson filter, final List<? extends Bson> update, final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateMany(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -730,13 +730,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<UpdateResult> updateMany(final ClientSession clientSession, final Bson filter, final List<? extends Bson> update,
                                               final UpdateOptions options) {
-        return new SingleResultObservableToPublisher<UpdateResult>(
+        return new ObservableToPublisher<UpdateResult>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<UpdateResult>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<UpdateResult> callback) {
                         wrapped.updateMany(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -746,13 +746,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<TDocument> findOneAndDelete(final Bson filter, final FindOneAndDeleteOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndDelete(filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -763,13 +763,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<TDocument> findOneAndDelete(final ClientSession clientSession, final Bson filter,
                                                  final FindOneAndDeleteOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndDelete(clientSession.getWrapped(), filter, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -779,13 +779,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<TDocument> findOneAndReplace(final Bson filter, final TDocument replacement, final FindOneAndReplaceOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndReplace(filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -796,13 +796,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<TDocument> findOneAndReplace(final ClientSession clientSession, final Bson filter, final TDocument replacement,
                                                   final FindOneAndReplaceOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndReplace(clientSession.getWrapped(), filter, replacement, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -812,13 +812,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<TDocument> findOneAndUpdate(final Bson filter, final Bson update, final FindOneAndUpdateOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndUpdate(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -829,13 +829,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<TDocument> findOneAndUpdate(final ClientSession clientSession, final Bson filter, final Bson update,
                                                  final FindOneAndUpdateOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndUpdate(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -846,13 +846,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<TDocument> findOneAndUpdate(final Bson filter, final List<? extends Bson> update,
                                                  final FindOneAndUpdateOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndUpdate(filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -864,35 +864,35 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<TDocument> findOneAndUpdate(final ClientSession clientSession, final Bson filter,
                                                  final List<? extends Bson> update, final FindOneAndUpdateOptions options) {
-        return new SingleResultObservableToPublisher<TDocument>(
+        return new ObservableToPublisher<TDocument>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<TDocument>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<TDocument> callback) {
                         wrapped.findOneAndUpdate(clientSession.getWrapped(), filter, update, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> drop() {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.drop(voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> drop(final ClientSession clientSession) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.drop(clientSession.getWrapped(), voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -902,13 +902,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<String> createIndex(final Bson key, final IndexOptions options) {
-        return new SingleResultObservableToPublisher<String>(
+        return new ObservableToPublisher<String>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<String>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<String> callback) {
                         wrapped.createIndex(key, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -918,13 +918,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<String> createIndex(final ClientSession clientSession, final Bson key, final IndexOptions options) {
-        return new SingleResultObservableToPublisher<String>(
+        return new ObservableToPublisher<String>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<String>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<String> callback) {
                         wrapped.createIndex(clientSession.getWrapped(), key, options, callback);
                     }
-                });
+                }));
     }
 
     @Override
@@ -992,24 +992,24 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Success> dropIndex(final String indexName, final DropIndexOptions dropIndexOptions) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.dropIndex(indexName, dropIndexOptions, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> dropIndex(final Bson keys, final DropIndexOptions dropIndexOptions) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.dropIndex(keys, dropIndexOptions, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -1025,24 +1025,24 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<Success> dropIndex(final ClientSession clientSession, final String indexName,
                                         final DropIndexOptions dropIndexOptions) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.dropIndex(clientSession.getWrapped(), indexName, dropIndexOptions, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
     public Publisher<Success> dropIndex(final ClientSession clientSession, final Bson keys, final DropIndexOptions dropIndexOptions) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.dropIndex(clientSession.getWrapped(), keys, dropIndexOptions, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -1072,13 +1072,13 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
 
     @Override
     public Publisher<Success> renameCollection(final MongoNamespace newCollectionNamespace, final RenameCollectionOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.renameCollection(newCollectionNamespace, options, voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
     @Override
@@ -1089,14 +1089,14 @@ final class MongoCollectionImpl<TDocument> implements MongoCollection<TDocument>
     @Override
     public Publisher<Success> renameCollection(final ClientSession clientSession, final MongoNamespace newCollectionNamespace,
                                                final RenameCollectionOptions options) {
-        return new SingleResultObservableToPublisher<Success>(
+        return new ObservableToPublisher<Success>(com.mongodb.async.client.Observables.observe(
                 new Block<com.mongodb.async.SingleResultCallback<Success>>() {
                     @Override
                     public void apply(final com.mongodb.async.SingleResultCallback<Success> callback) {
                         wrapped.renameCollection(clientSession.getWrapped(), newCollectionNamespace, options,
                                 voidToSuccessCallback(callback));
                     }
-                });
+                }));
     }
 
 }
